Feature/security signup

+++ backend/src/main/java/com/github/edona94/service/EmployeeService.java
@@ -1,6 +1,7 @@
 package com.github.edona94.service;
 
 import com.github.edona94.exception.EmployeeNotFoundException;
+import com.github.edona94.exception.UnauthorizedException;
 import com.github.edona94.model.Employee;
 import com.github.edona94.model.EmployeeDTORequest;
 import com.github.edona94.repository.EmployeeRepository;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.security.Principal;
 import java.util.InputMismatchException;
 import java.util.List;
 import java.util.Optional;
@@ -19,9 +21,10 @@ public class EmployeeService {
     private final EmployeeRepository employeeRepository;
     private final IdService idService;
     private final CVService cvService;
+    private final MongoUserDetailsService mongoUserDetailsService;
 
     public List<Employee> getAllEmployees(){
-        return employeeRepository.findAll();
+        return employeeRepository.findAllByOrderByAddedDesc();
     }
 
     public String uploadCv(MultipartFile cv) {
@@ -38,9 +41,10 @@ public class EmployeeService {
         return cvUri;
     }
 
-    public Employee addEmployee(EmployeeDTORequest employeeDTORequest, MultipartFile cv) {
+    public Employee addEmployee(EmployeeDTORequest employeeDTORequest, MultipartFile cv, Principal principal) {
         String id = idService.generateId();
         String cvUri = uploadCv(cv);
+        String adminId = mongoUserDetailsService.getMe(principal).id();
         Employee newEmployee = new Employee(
                 id,
                 employeeDTORequest.firstName(),
@@ -51,7 +55,8 @@ public class EmployeeService {
                 employeeDTORequest.email(),
                 employeeDTORequest.phoneNumber(),
                 employeeDTORequest.added(),
-                cvUri
+                cvUri,
+                adminId
         );
         return employeeRepository.save(newEmployee);
     }
@@ -60,14 +65,17 @@ public class EmployeeService {
         return employeeRepository.findById(id).orElseThrow(EmployeeNotFoundException::new);
     }
 
-    public Employee updateEmployeeById(String id, EmployeeDTORequest employeeDTORequest, MultipartFile cv) {
+    public Employee updateEmployeeById(String id, EmployeeDTORequest employeeDTORequest, MultipartFile cv,Principal principal) {
+        String adminId = mongoUserDetailsService.getMe(principal).id();
         Employee employee = getEmployeeById(id);
         String cvUri = employee.cv(); // Set the cvUri to the existing CV URI by default
 
         if(cv != null && !cv.isEmpty()) { // Check if a new CV file has been provided
             cvUri = uploadCv(cv); // If yes, upload the new CV and update the cvUri
         }
-
+        if (!employee.userId().equals(adminId)) {
+            throw new UnauthorizedException("Only Admin can edit the employee data");
+        }
         Employee updatedEmployee = new Employee(
                 id,
                 employeeDTORequest.firstName(),
@@ -78,15 +86,20 @@ public class EmployeeService {
                 employeeDTORequest.email(),
                 employeeDTORequest.phoneNumber(),
                 employeeDTORequest.added(),
-                cvUri
+                cvUri,
+                adminId
         );
         return employeeRepository.save(updatedEmployee);
     }
 
-    public Employee deleteEmployee(String id) {
+    public Employee deleteEmployee(String id,Principal principal) {
+        String adminId = mongoUserDetailsService.getMe(principal).id();
         Optional<Employee> employee = employeeRepository.findById(id);
         if(employee.isEmpty()) {
             throw new EmployeeNotFoundException("Employee with id" +id+ "doesn't exist");
+        }
+        if(!employee.get().userId().equals(adminId)){
+            throw new UnauthorizedException("Only Admin can delete an employee");
         }
         employeeRepository.deleteById(id);
         return employee.get();
